requirements.txt
torch>=1.12.0
torch-geometric>=2.1.0
pytorch-lightning>=1.7.0
transformers>=4.20.0
datasets>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0
tensorboard>=2.8.0
pyyaml>=6.0
networkx>=2.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.12.0
omegaconf>=2.1.0
hydra-core>=1.1.0
